ibverbs engine unit tests fails (#4)
This MR fixes the following issues:
 1) -fsanitize=address detected one shared_ptr was not given the correct destructor on reset, herewith fixed;
 2) if ibverbs driver calls returns `NULL`, this is not of the C++ type `nullptr_t` and thus the shared-pointer reset did not operate as expected (it did not become empty), herewith fixed;
 3) some previously unlogged errors are now logged;
 4) trailing spaces removed throughout (code style).

Ad item no. 2: this error became visible when executing on a machine with a faulty ibverbs driver version that misrepresents the maximum number of outstanding RDMA requests supported. Running the `manyPuts` ibverbs test then resulted in a segfault, initially reported on Gitee as issue I515SM, herewith fixed.

The issue was a coding pattern where reset on a shared_ptr was assumed to set the shared_ptr to empty, when ibverbs calls returned `NULL`. This pattern occurred elsewhere also, beyond the point that resulted in a segfault in the above-mentioned issue. This MR applies the fix to all such cases; i.e., if `NULL` is returned, it now explicitly resets the shared_ptr to empty, which will then ensure the expected error handling paths fire, resulting in both logging of an appropriate error message, as well as in the throwing of an appropriate exception.